+++ st_link_analysis/frontend/src/components/infopanel.js
@@ -1,11 +1,12 @@
 import State from "../utils/state";
 
 // Constants / Configurations
-const INFOBAR_ID = "infobar";
-const LABEL_ID = "infobarLabel";
-const PROPS_ID = "infobarProps";
+const INFOPANEL_ID = "infopanel";
+const LABEL_ID = "infopanelLabel";
+const PROPS_ID = "infopanelProps";
+const NODEACTIONS_ID = "nodeActions";
 
-// Infobar children updates
+// Infopanel children updates
 function _updateLabel(color, label, icon) {
     const label_div = document.getElementById(LABEL_ID);
     label_div.firstChild.innerText = label;
@@ -26,35 +27,37 @@ function _updateProps(data) {
         })
         .map(([key, value]) => {
             return `
-        <div class='infobar__prop'>
-            <p class='infobar__key'>${key}</p>
-            <p class='infobar__val'>${value}</p>
+        <div class='infopanel__prop'>
+            <p class='infopanel__key'>${key}</p>
+            <p class='infopanel__val'>${value}</p>
         </div>`;
         })
         .join("");
 }
 
-// Infobar update
-function updateInfobar() {
-    const infobar = document.getElementById(INFOBAR_ID);
+// infopanel update
+function updateInfopanel() {
+    const infopanel = document.getElementById(INFOPANEL_ID);
+    const nodeActions = document.getElementById(NODEACTIONS_ID);
     const el = State.getState("selection").lastSelected;
-    let color, data, label, selected, icon;
+    let color, data, label, expanded, icon;
     if (el) {
         color = el.style().backgroundColor;
         data = el.data();
         label = data["label"] || el.group().slice(0, -1).toUpperCase();
-        selected = "element";
+        expanded = true;
         icon = el.style()["background-image"];
     } else {
         color = "hsla(0, 0%, 0%, 0)";
         data = {};
         label = "";
-        selected = null;
+        expanded = false;
         icon = null;
     }
-    infobar.setAttribute("data-selected", selected);
+    infopanel.setAttribute("data-expanded", expanded);
+    nodeActions.setAttribute("data-expanded", expanded);
     _updateLabel(color, label, icon);
     _updateProps(data);
 }
 
-export default updateInfobar;
+export default updateInfopanel;