+++ tests/test_node_actions.py
@@ -1,10 +1,10 @@
 from playwright.sync_api import Page
 import json
+import re
 
 
-PAGE_NAME = "Events Listeners"
-NODE_ID = "n4"
-EDGE_ID = "e2"
+PAGE_NAME = "Node Actions"
+NODE_ID = "c"
 ASSIGN_CY = "const cy = document.getElementById('cy')._cyreg.cy;"
 FRAME_LOCATOR = "iframe[title*='st_link_analysis']"
 
@@ -17,14 +17,6 @@ def get_node_pos(_id, iframe):
     return pos
 
 
-def get_edge_pos(_id, iframe):
-    pos = iframe.evaluate(f"""() => {{
-        {ASSIGN_CY}
-        return cy.getElementById("{_id}").renderedMidpoint();
-    }}""")
-    return pos
-
-
 def get_return_json(page: Page):
     data = (
         page.get_by_test_id("stJson")
@@ -32,6 +24,7 @@ def get_return_json(page: Page):
         .replace('""', '","')
         .replace('}"', '},"')
     )
+    data = re.sub("([0-9]+):", "", data)
     return json.loads(data)
 
 
@@ -42,41 +35,59 @@ def test_iframe_exists_events(page: Page):
     assert len(frames) == 1
 
 
-def test_single_click_node_event(page: Page):
+def test_expand_dblclick(page: Page):
+    page.get_by_role("link", name=PAGE_NAME).click()
+    frame = page.frame_locator(FRAME_LOCATOR).first.locator(":root")
+    frame.click(position={"x": 0, "y": 0})  # await and scroll to view
+
+    pos = get_node_pos(NODE_ID, frame)
+    frame.dblclick(position=pos)
+    page.wait_for_timeout(500)
+    data = get_return_json(page)
+    assert data["action"] == "expand"
+    assert data["data"]["node_ids"][0] == NODE_ID
+
+
+def test_expand_button(page: Page):
     page.get_by_role("link", name=PAGE_NAME).click()
     frame = page.frame_locator(FRAME_LOCATOR).first.locator(":root")
     frame.click(position={"x": 0, "y": 0})  # await and scroll to view
 
     pos = get_node_pos(NODE_ID, frame)
     frame.click(position=pos)
-    page.wait_for_timeout(300)
+    page.wait_for_timeout(200)
+    frame.get_by_title("Expand Node").click()
+    page.wait_for_timeout(500)
     data = get_return_json(page)
-    assert data["data"]["target_id"] == NODE_ID
-    assert data["data"]["target_group"] == "nodes"
-    assert data["action"] == "clicked_node"
+    assert data["action"] == "expand"
+    assert data["data"]["node_ids"][0] == NODE_ID
 
 
-def test_double_click_edge_event(page: Page):
+def test_remove_keydown(page: Page):
     page.get_by_role("link", name=PAGE_NAME).click()
     frame = page.frame_locator(FRAME_LOCATOR).first.locator(":root")
     frame.click(position={"x": 0, "y": 0})  # await and scroll to view
 
-    pos = get_edge_pos(EDGE_ID, frame)
-    frame.dblclick(position=pos)
-    page.wait_for_timeout(300)
-    data = get_return_json(page)
-    assert data["data"]["target_id"] == EDGE_ID
-    assert data["data"]["target_group"] == "edges"
-    assert data["action"] == "another_name"
-
-
-def test_single_click_edge_no_event(page: Page):
-    page.get_by_role("link", name=PAGE_NAME).click()
-    frame = page.frame_locator(FRAME_LOCATOR).first.locator(":root")
-    frame.click(position={"x": 0, "y": 0})  # await and scroll to view
-
-    pos = get_edge_pos(EDGE_ID, frame)
+    pos = get_node_pos(NODE_ID, frame)
     frame.click(position=pos)
-    page.wait_for_timeout(250)
+    page.wait_for_timeout(200)
+    page.keyboard.down("Delete")
+    page.wait_for_timeout(500)
     data = get_return_json(page)
-    assert data == {}
+    assert data["action"] == "remove"
+    assert data["data"]["node_ids"][0] == NODE_ID
+
+
+def test_remove_button(page: Page):
+    page.get_by_role("link", name=PAGE_NAME).click()
+    frame = page.frame_locator(FRAME_LOCATOR).first.locator(":root")
+    frame.click(position={"x": 0, "y": 0})  # await and scroll to view
+
+    pos = get_node_pos(NODE_ID, frame)
+    frame.click(position=pos)
+    page.wait_for_timeout(200)
+    frame.get_by_title("Remove Nodes").click()
+    page.wait_for_timeout(500)
+    data = get_return_json(page)
+    assert data["action"] == "remove"
+    assert data["data"]["node_ids"][0] == NODE_ID